requirements.txt
google-generativeai==0.3.1
streamlit==1.29.0
joblib==1.3.2
langchain-google-genai
langchain
langchain-community
pathlib
PyPDF2
easyocr
pyodbc
streamlit_option_menu
python-dotenv
pdf2image
matplotlib